restructure the project to meet webservice
add spring-boot-starter-web dependency to the pom
move the class Measurement and class Measurements deom model package to Business.domain
move the parser classes to business.service
