+++ src/main/java/org/n52/datareader/business/domain/Measurements.java
@@ -1,4 +1,4 @@
-package org.n52.datareader.model;
+package org.n52.datareader.business.domain;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
