+++ src/main/java/org/n52/datareader/business/service/DateFormatter.java
@@ -1,10 +1,18 @@
-package org.n52.datareader.service;
+package org.n52.datareader.business.service;
 
 import javax.xml.bind.annotation.adapters.XmlAdapter;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
-
+/* xml adapeter is used in JAXB and to be automatically adjusted
+*    @XmlJavaTypeAdapter(DateFormatter.class)
+    @JsonProperty(TIME)
+    @XmlElement(name = TIME)
+    public void setTime(Date time) {
+        this.time = time;
+    }
+*
+*  */
 public class DateFormatter extends XmlAdapter<String, Date> {
 
    public static final String CUSTOM_FORMAT_STRING = "yyyy-MM-dd'T'HH:mm:ss'Z'";
